Rename `package hostprofilerrunner` to `runner`
