+++ main.go
@@ -18,26 +18,26 @@ import (
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/sys/unix"
 
-	"github.com/DataDog/dd-otel-host-profiler/hostprofilerrunner"
+	"github.com/DataDog/dd-otel-host-profiler/runner"
 )
 
 func main() {
 	os.Exit(int(mainWithExitCode()))
 }
 
-func mainWithExitCode() hostprofilerrunner.ExitCode {
-	args, err := hostprofilerrunner.ParseArgs()
+func mainWithExitCode() runner.ExitCode {
+	args, err := runner.ParseArgs()
 	if err != nil {
-		return hostprofilerrunner.ParseError("Failure to parse arguments: %v", err)
+		return runner.ParseError("Failure to parse arguments: %v", err)
 	}
 
 	if args == nil {
-		return hostprofilerrunner.ExitSuccess
+		return runner.ExitSuccess
 	}
 
 	if args.Copyright {
-		fmt.Print(hostprofilerrunner.Copyright)
-		return hostprofilerrunner.ExitSuccess
+		fmt.Print(runner.Copyright)
+		return runner.ExitSuccess
 	}
 
 	// Context to drive main goroutine and the Tracer monitors.
@@ -51,5 +51,5 @@ func mainWithExitCode() hostprofilerrunner.ExitCode {
 		args.Dump()
 	}
 
-	return hostprofilerrunner.RunHostProfiler(mainCtx, &args.FullHostProfilerSettings)
+	return runner.RunHostProfiler(mainCtx, &args.FullHostProfilerSettings)
 }
