+++ runner/fullhostprofilersettings.go
@@ -1,7 +1,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024 Datadog, Inc.
 
-package hostprofilerrunner
+package runner
 
 import (
 	"encoding/json"
