+++ runner/helpers.go
@@ -7,7 +7,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024 Datadog, Inc.
 
-package hostprofilerrunner
+package runner
 
 import (
 	"fmt"
