+++ runner/helpers_test.go
@@ -1,7 +1,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025 Datadog, Inc.
 
-package hostprofilerrunner
+package runner
 
 import (
 	"reflect"
